pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "infrakit"
version = "0.1.0"
description = "Declarative AWS infrastructure deployment from a single YAML file"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Samuel Okechukwu", email = "infrakit@sokech.com" }]
keywords = ["aws", "infrastructure", "iac", "devops", "cli", "cloudformation", "cdk"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Intended Audience :: System Administrators",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Build Tools",
  "Topic :: System :: Systems Administration",
]

dependencies = [
  "typer[all]>=0.12.0",         # CLI framework (wraps Click, adds Rich)
  "pydantic>=2.0.0",            # Schema validation
  "pyyaml>=6.0",                # YAML parsing
  "boto3>=1.34.0",              # AWS SDK
  "rich>=13.0.0",               # Terminal output (tables, progress, color)
  "networkx>=3.0",              # Dependency DAG resolution
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "moto[all]>=5.0",             # AWS API mocking
  "ruff>=0.4.0",                # Linter + formatter
  "mypy>=1.10",
  "boto3-stubs[essential]>=1.34.0",  # Type stubs for boto3
  "pre-commit>=3.7",
]

[project.scripts]
infrakit = "infrakit.cli.main:app"

[project.urls]
Homepage = "https://github.com/SamioneX/InfraKit"
Repository = "https://github.com/SamioneX/InfraKit"
Documentation = "https://github.com/SamioneX/InfraKit#readme"
"Bug Tracker" = "https://github.com/SamioneX/InfraKit/issues"

# ── Hatch build config ────────────────────────────────────────────────────────
[tool.hatch.build.targets.wheel]
packages = ["src/infrakit"]

# ── Ruff (linter + formatter) ─────────────────────────────────────────────────
[tool.ruff]
src = ["src"]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "TCH"]
ignore = ["E501"]  # line length handled by formatter

[tool.ruff.lint.isort]
known-first-party = ["infrakit"]

# ── Mypy ──────────────────────────────────────────────────────────────────────
[tool.mypy]
strict = true
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

# ── Pytest ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src/infrakit --cov-report=term-missing --cov-fail-under=90"
markers = [
  "unit: fast unit tests, no network",
  "integration: uses moto for AWS mocking",
]

# ── Coverage ──────────────────────────────────────────────────────────────────
[tool.coverage.run]
source = ["src/infrakit"]
omit = ["*/tests/*", "*/conftest.py"]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "raise NotImplementedError",
  "if TYPE_CHECKING:",
]
